\name{cartcpp-package}
\alias{cartcpp-package}
\alias{cartcpp}
\docType{package}
\title{
\packageTitle{cartcpp}
}
\description{
\packageDescription{cartcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cartcpp}
\packageIndices{cartcpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cartcpp}

Maintainer: \packageMaintainer{cartcpp}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
