% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree$new}
\alias{Tree$new}
\title{Constructs a new Tree object}
\arguments{
\item{ident}{ID of the tree, must be an integer number}

\item{treeType}{0 for classification tree and 1 for regresssion tree}

\item{maxNumFeatures}{Number of features in the data set on which the model is trained}

\item{numFeatures}{Number of features selected at each split}

\item{maxDepth}{The maximum depth to which the tree is grown}

\item{minCount}{Minimum number of data points for a node to qualify as a leaf node}
}
\description{
Constructs a new Tree object
}
\examples{
# Create a new object of class Tree with the given parameters
tr = new(Tree, ident = 123, treeType = 0, maxNumFeatures = 20,
numFeatures = 3, maxDepth = 4, minCount = 2)
}
