% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree$predict}
\alias{Tree$predict}
\title{Calculates predictions based on the Tree model}
\arguments{
\item{X}{Data matrix}
}
\value{
Vector of predictions corresponding to the data.
}
\description{
Calculates predictions based on the Tree model
}
\examples{
# Define a tree object
tr = new(Tree, ident = 123, treeType = 0, maxNumFeatures = 4,
numFeatures = 3, maxDepth = 10, minCount = 2)
# Create a training set
X = matrix(c(rnorm(16, 0, 2), rnorm(16, 20, 2)), nrow = 8, ncol = 4, byrow = TRUE)
Y = c(0, 0, 1, 1, 2, 2, 2, 2)
# Train the model
tr$train(X, Y)
Define the test set
Xtest = matrix(c(rnorm(8, 0, 2), rnorm(8, 20, 2)), nrow = 4, ncol = 4, byrow = TRUE)
# Calculate predictions vector
tr$predict(Xtest)
}
