% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree$print}
\alias{Tree$print}
\title{Prints the tree structure in the matrix form.}
\value{
The data matrix, containing the description of the structure of the tree model. The first
column indicates the depth of the node. The second column indicates whether the node is a leaf: 0 - not a leaf
1 - leaf. The third column identifies the feature index based on which the split was performed. The forth column
gives the splitting value: all data points which have the feature value less than or equal (<=) to the splitting
value are mapped to the left node. Finally, the fifth column indicates the leaf node value if the node is a leaf.
}
\description{
The consecutive rows of matrix represent the nodes. The matrix is formed as follows: first, the node
is printed, then the recursive calls are made to print
its left and right child nodes respectively. Due to the recursive nature of the print function, the matrix
representing the structure of the tree must be handled with care and requires proper attention.
}
\examples{
# Assume a trained and defined Tree object, tr
tr.Print()
}
