% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree$train}
\alias{Tree$train}
\title{Fits a Tree object to the given data}
\arguments{
\item{X}{Data matrix}

\item{Y}{Vector of labels}
}
\description{
Fits a Tree object to the given data
}
\examples{
# Define a Tree object
tr = new(Tree, ident = 123, treeType = 0, maxNumFeatures = 4,
numFeatures = 3, maxDepth = 10, minCount = 2)
# Create a training set
X = matrix(c(rnorm(16, 0, 2), rnorm(16, 20, 2)), nrow = 8, ncol = 4, byrow = TRUE)
Y = c(0, 0, 1, 1, 2, 2, 2, 2)
# Train the model
tr$train(X, Y)
}
